async finished

+++ src/main/kotlin/com/grispi/grispiimport/zendesk/ticket/ZendeskComments.kt
@@ -61,7 +61,7 @@ class ZendeskComment: ZendeskEntity() {
         val ticketKey = getGrispiTicketKey.invoke(ticketId!!)
         val grispiUserId = getGrispiUserId.invoke(authorId)
         val attachmentRequests = attachments.map { AttachmentRequest(it.url.toString(), it.filename.toString()) }
-        return CommentRequest(EmojiParser.parseToAliases(htmlBody), public, ticketKey, createdAt.toString(), CommentCreator(grispiUserId.toLong(), null, null), attachmentRequests)
+        return CommentRequest(EmojiParser.parseToAliases(htmlBody), public, ticketKey, createdAt.toString(), CommentCreator(grispiUserId, null, null), attachmentRequests)
     }
 }
